+++ src/watchdog.c
@@ -9,7 +9,7 @@
 #define WATCHDOG_SLEEP 2000 / N_ROUNDS
 
 void* watchdog(void* arg) {
-    Warehouse* w = *(Warehouse**) arg;
+    Warehouse* w = *(Warehouse**)arg;
     int reader_tmp;
     int analyser_tmp;
     int printer_tmp;
@@ -24,19 +24,19 @@ void* watchdog(void* arg) {
 
         reader_tmp = warehouse_watchdog_check_reader(w);
         if(reader_tmp != -1) {
-            reader_cnt = (reader_cnt + 1) * reader_tmp;
+            reader_cnt = (reader_cnt + 1) * (size_t)reader_tmp;
         }
         analyser_tmp = warehouse_watchdog_check_analyser(w);
         if(analyser_tmp != -1) {
-            analyser_cnt = (analyser_cnt + 1) * analyser_tmp;
+            analyser_cnt = (analyser_cnt + 1) * (size_t)analyser_tmp;
         }
         printer_tmp = warehouse_watchdog_check_printer(w);
         if(printer_tmp != -1) {
-            printer_cnt = (printer_cnt + 1) * printer_tmp;
+            printer_cnt = (printer_cnt + 1) * (size_t)printer_tmp;
         }
         logger_tmp = warehouse_watchdog_check_logger(w);
         if(logger_tmp != -1) {
-            logger_cnt = (logger_cnt + 1) * logger_tmp;
+            logger_cnt = (logger_cnt + 1) * (size_t)logger_tmp;
         }
     }
